Refactor plugin structure under src
