+++ composer.json
@@ -8,8 +8,13 @@
         "phpoffice/phpspreadsheet": "1.30"
     },
     "config": {
-    "platform": {
-        "php": "8.2.27"
+        "platform": {
+            "php": "8.2.27"
+        }
+    },
+    "autoload": {
+        "psr-4": {
+            "TutoriasBooking\\": "src/"
+        }
     }
-}
 }
