+++ src/Admin/AdminPage.php
@@ -1,8 +1,6 @@
 <?php
 namespace TutoriasBooking\Admin;
 
-require_once TB_PLUGIN_DIR . 'includes/Admin/AdminController.php';
-
 /**
  * Wrapper function to display the admin page.
  */
